requirements.txt
streamlit
pandas
numpy
finlab>=0.3.11
plotly
yfinance
scipy
google-genai
graphviz
requests
shioaji
python-dateutil
